Applying minimal changes necessary to get config entries working.
