+++ homeassistant/components/vera/switch.py
@@ -12,7 +12,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import Entity
 from homeassistant.util import convert
 
-from .common import VeraDevice, setup_device_entities
+from . import VeraDevice
+from .common import setup_device_entities
 
 _LOGGER = logging.getLogger(__name__)
