+++ tests/components/vera/conftest.py
@@ -9,5 +9,5 @@ from .common import ComponentFactory
 @pytest.fixture()
 def vera_component_factory():
     """Return a factory for initializing the vera component."""
-    with patch("pyvera.VeraController") as vera_controller_class_mock:
-        yield ComponentFactory(vera_controller_class_mock)
+    with patch("pyvera.init_controller") as init_controller_mock:
+        yield ComponentFactory(init_controller_mock)
